Clean up Hamiltonian classes

+++ src/quant_met/geometry/base_lattice.py
@@ -19,7 +19,7 @@ class BaseLattice(ABC):
 
     @property
     @abstractmethod
-    def lattice_constant(self) -> np.float64:
+    def lattice_constant(self) -> float:
         """Lattice constant."""
         raise NotImplementedError
