+++ src/quant_met/geometry/graphene.py
@@ -13,7 +13,7 @@ from .base_lattice import BaseLattice
 class GrapheneLattice(BaseLattice):
     """Lattice geometry for Graphene."""
 
-    def __init__(self, lattice_constant: np.float64) -> None:
+    def __init__(self, lattice_constant: float) -> None:
         self._lattice_constant = lattice_constant
         self._bz_corners = (
             4
@@ -27,7 +27,7 @@ class GrapheneLattice(BaseLattice):
         self._high_symmetry_points = ((self.M, "M"), (self.Gamma, r"\Gamma"), (self.K, "K"))
 
     @property
-    def lattice_constant(self) -> np.float64:  # noqa: D102
+    def lattice_constant(self) -> float:  # noqa: D102
         return self._lattice_constant
 
     @property
